requirements/common.txt
requests>=2.25.1
websocket-client>=1.5.0
pycryptodome>=3.15.0
python-dotenv>=1.0.1
